requirements.txt
stravalib==1.5
pandas
python-dotenv
fastparquet
pyarrow
scikit-learn
tensorflow
joblib
flask
gunicorn
Flask-SQLAlchemy
pytest
pytest-cov
requests-mock
